requirements.txt
streamlit
google-generativeai
python-dotenv
scikit-learn
pandas
numpy
matplotlib